Refactor: Move dedup logic to deduplication file
